La til ny consumer client for å rullere aiven credentials
