+++ kafka/src/main/java/no/nav/common/kafka/consumer/KafkaConsumerClientConfig.java
@@ -2,11 +2,13 @@ package no.nav.common.kafka.consumer;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import no.nav.common.kafka.consumer.util.ConsumerClientExceptionListener;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 
-import static no.nav.common.kafka.consumer.KafkaConsumerClient.DEFAULT_POLL_DURATION_MS;
+import static no.nav.common.kafka.consumer.KafkaConsumerClientImpl.DEFAULT_POLL_DURATION_MS;
 
 @AllArgsConstructor
 @Data
@@ -17,6 +19,8 @@ public class KafkaConsumerClientConfig<K, V> {
     Map<String, TopicConsumer<K, V>> topics;
 
     long pollDurationMs;
+
+    List<ConsumerClientExceptionListener> exceptionListeners;
 
     public KafkaConsumerClientConfig(Properties properties, Map<String, TopicConsumer<K, V>> topics) {
         this.properties = properties;
