+++ kafka/src/main/java/no/nav/common/kafka/consumer/util/ConsumerClientExceptionListener.java
@@ -1,0 +1,7 @@
+package no.nav.common.kafka.consumer.util;
+
+public interface ConsumerClientExceptionListener {
+
+    void onExceptionCaught(Exception e);
+
+}
